[Redis] feat: move redis 파일 이동
