--- src/main/java/com/mzc/configuration/RedisConfiguration.java
@@ -1,54 +1,0 @@
-package com.mzc.configuration;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.RedisMessageListenerContainer;
-import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
-import org.springframework.data.redis.serializer.GenericToStringSerializer;
-import org.springframework.stereotype.Component;
-
-@Component
-@Configuration
-public class RedisConfiguration {
-
-    @Autowired
-    private RedisConnectionFactory redisConnectionFactory;
-
-    @Bean
-    public RedisMessageListenerContainer container(RedisConnectionFactory redisConnectionFactory, MessageListenerAdapter messageListenerAdapter){
-        RedisMessageListenerContainer redisMessageListenerContainer = new RedisMessageListenerContainer();
-        redisMessageListenerContainer.setConnectionFactory(redisConnectionFactory);
-        redisMessageListenerContainer.addMessageListener(messageListenerAdapter,topic());
-        return redisMessageListenerContainer;
-    }
-
-    @Bean
-    MessageListenerAdapter messageListenerAdapter()
-    {
-        return new MessageListenerAdapter(new RedisMessageSubscriber(),"onMessage");
-    }
-    @Bean
-    public ChannelTopic topic() {
-        return new ChannelTopic("RedisQuiz");
-    }
-
-    @Bean
-    RedisTemplate<String,Object> redisTemplate(RedisConnectionFactory redisConnectionFactory)
-    {
-        RedisTemplate<String,Object> redisTemplate = new RedisTemplate<>();
-        redisTemplate.setConnectionFactory(redisConnectionFactory);
-        redisTemplate.setValueSerializer(new GenericToStringSerializer<Object>(Object.class));
-        return redisTemplate;
-    }
-
-    @Bean
-    MessagePublisher messagePublisher()
-    {
-        return new RedisMessagePublisher(redisTemplate(redisConnectionFactory),topic());
-    }
-
-}
