+++ src/main/java/com/mzc/global/config/RedisConfig.java
@@ -1,58 +1,57 @@
-//package com.mzc.global.config;
-//
-//import io.lettuce.core.RedisURI;
-//import lombok.RequiredArgsConstructor;
-//import org.springframework.boot.autoconfigure.data.redis.RedisProperties;
-//import org.springframework.context.annotation.Bean;
-//import org.springframework.context.annotation.Configuration;
-//import org.springframework.data.redis.connection.RedisConnection;
-//import org.springframework.data.redis.connection.RedisConnectionFactory;
-//import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-//import org.springframework.data.redis.core.RedisTemplate;
-//import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
-//import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
-//import org.springframework.data.redis.serializer.StringRedisSerializer;
-//
-//@RequiredArgsConstructor
-//@Configuration
-//@EnableRedisRepositories
-//public class RedisConfig {
-//    private final RedisProperties redisProperties;
-//
-////    @Bean
-////    public RedisConnectionFactory redisConnectionFactory(){
-////        return new LettuceConnectionFactory(redisProperties.getHost(), redisProperties.getPort());
-////    }
-//
-//    //@Bean
-//    public RedisTemplate<?, ?> redisTemplate(){
-//        RedisTemplate<byte[], byte[]> redisTemplate = new RedisTemplate<>();
-//        redisTemplate.setConnectionFactory(redisConnectionFactory());
-//        redisTemplate.setKeySerializer(new StringRedisSerializer());
-//        return redisTemplate;
-//    }
-//
-//   // @Bean
-//    public RedisConnectionFactory redisConnectionFactory(){
-//        RedisURI redisHost = RedisURI.create(redisProperties.getHost());
-//        RedisURI redisPort = RedisURI.create(String.valueOf(redisProperties.getPort()));
-//
-//        org.springframework.data.redis.connection.RedisConfiguration configuration = LettuceConnectionFactory.createRedisConfiguration(redisHost); // 제디스, 레투스 방법 으로 연결
-//        LettuceConnectionFactory factory = new LettuceConnectionFactory(configuration);
-//        factory.afterPropertiesSet();
-//        return factory;
-//    }
-//
-//    // 접근이 자주 되는 데이터를 캐싱하는 것이 좋음
-//   //Bean
-///*    public RedisTemplate<String, User> userRedisTemplate(RedisConnectionFactory redisConnectionFactory){
-//        RedisTemplate<String, User> redisTemplate = new RedisTemplate<>();
-//        redisTemplate.setConnectionFactory(redisConnectionFactory);
-//        //redisTemplate.opsForValue(); // get/set 기본적으로 할 수 있게 해주는 것
-//        redisTemplate.setKeySerializer(new StringRedisSerializer());
-//        redisTemplate.setValueSerializer(new Jackson2JsonRedisSerializer<User>(User.class));
-//        return redisTemplate;
-//    }*/
-//
-//
-//}
+package com.mzc.global.config;
+
+import com.mzc.redis.pub.MessagePublisher;
+import com.mzc.redis.pub.RedisMessagePublisher;
+import com.mzc.redis.sub.RedisMessageSubscriber;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.listener.ChannelTopic;
+import org.springframework.data.redis.listener.RedisMessageListenerContainer;
+import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
+import org.springframework.data.redis.serializer.GenericToStringSerializer;
+import org.springframework.stereotype.Component;
+
+@Component
+@Configuration
+public class RedisConfig {
+
+    @Autowired
+    private RedisConnectionFactory redisConnectionFactory;
+
+    @Bean
+    public RedisMessageListenerContainer container(RedisConnectionFactory redisConnectionFactory, MessageListenerAdapter messageListenerAdapter){
+        RedisMessageListenerContainer redisMessageListenerContainer = new RedisMessageListenerContainer();
+        redisMessageListenerContainer.setConnectionFactory(redisConnectionFactory);
+        redisMessageListenerContainer.addMessageListener(messageListenerAdapter,topic());
+        return redisMessageListenerContainer;
+    }
+
+    @Bean
+    MessageListenerAdapter messageListenerAdapter()
+    {
+        return new MessageListenerAdapter(new RedisMessageSubscriber(),"onMessage");
+    }
+    @Bean
+    public ChannelTopic topic() {
+        return new ChannelTopic("RedisQuiz");
+    }
+
+    @Bean
+    RedisTemplate<String,Object> redisTemplate(RedisConnectionFactory redisConnectionFactory)
+    {
+        RedisTemplate<String,Object> redisTemplate = new RedisTemplate<>();
+        redisTemplate.setConnectionFactory(redisConnectionFactory);
+        redisTemplate.setValueSerializer(new GenericToStringSerializer<Object>(Object.class));
+        return redisTemplate;
+    }
+
+    @Bean
+    MessagePublisher messagePublisher()
+    {
+        return new RedisMessagePublisher(redisTemplate(redisConnectionFactory),topic());
+    }
+
+}
