+++ src/main/java/com/mzc/redis/model/QuizMessage.java
@@ -1,12 +1,9 @@
-package com.mzc.model;
+package com.mzc.redis.model;
 
 import lombok.Data;
 
 @Data
-public class Message {
-
+public class QuizMessage {
     private String data;
     private String author;
-
-
 }
