+++ src/main/java/com/mzc/redis/pub/MessagePublisher.java
@@ -1,4 +1,4 @@
-package com.mzc.configuration;
+package com.mzc.redis.pub;
 
 public interface MessagePublisher {
     void publish(String message);
