+++ src/main/resources/application.properties
@@ -18,7 +18,7 @@ spring.data.mongodb.database=${env-mongodb-database}
 #keycloak.bearer-only=true
 
 #redis
-spring.redis.host=35.180.8.99
+spring.redis.host=3.36.65.78
 spring.redis.password=
 spring.redis.port=6379
 spring.redis.jedis.pool.max-idle=8
